Move state to its own header
